undo last changes

+++ main.h
@@ -40,7 +40,7 @@ int print_hexa(va_list args, char *buffer);
 int print_hexa_up(va_list args, char *buffer);
 int search_spec_to_print(const char *format, int i, int *counter,
 			 fspec *chartof, va_list args, char *buffer);
-char *store_base_reverse(size_t n, int base, int upper,
+char *store_base_reverse(unsigned int n, int base, int upper,
 			 char *buffer);
 int flush_buffer(char *buffer);
 int print_rot13(va_list args, char *buffer);
